Moved to Proper Folder. Double Checked DotEnv
